fix: restore issue workflow split dependencies
